test(i18n): centralise translation fixture loading (rung 2, commit 6/6)
Why: 8 *I18n-test.js suites each independently `require(...data.<locale>.json)`
and ran their own copy of `flattenMessages`. Webpack parsed each JSON up to
8 times. Move both into a shared module loaded once.

New fixture: js/__tests__/fixtures/translations.js
- Loads en-US, fr-FR, es-ES, ht-HT once (raw + flattened)
- Exports enData/frData/esData/htData, enMessages/frMessages/esMessages/
  htMessages, and flattenMessages (still used in-test by swammI18n which
  flattens additional locales inline)
- Filename intentionally has no `-test.js` suffix so it does not match the
  `js/**\/*-test.jsx?` glob in tests-travis.webpack.js

Path note: the static/ dir lives at geonode_mapstore_client/static/, NOT
under client/, so the fixture climbs 4 `../` (not 3 as a literal reading of
the implementation plan suggested). Verified each consumer's relative path
with `node -e "require.resolve(...)"` before running tests.

7 plugin tests at js/plugins/hydrata/<plugin>/__tests__/ are 5 levels deep
under client/ so they import via 4 `../`s. translationIntegrity-test.js at
js/plugins/hydrata/__tests__/ is 4 levels deep so it imports via 3 `../`s.

Verified: 1274 tests green; no behavioural change.

Per implementation plan:
docs/plans/2026-05-08-karma-rungs-1-2-implementation-prompt.md (commit 6).
