+++ geonode_mapstore_client/client/js/plugins/hydrata/SimpleView/__tests__/simpleViewI18n-test.js
@@ -4,24 +4,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import Localized from '@mapstore/framework/components/I18N/Localized';
 
-const enData = require('../../../../../../static/mapstore/hydrata-translations/data.en-US.json');
-const frData = require('../../../../../../static/mapstore/hydrata-translations/data.fr-FR.json');
-
-function flattenMessages(obj, prefix) {
-    let result = {};
-    Object.keys(obj).forEach(key => {
-        const fullKey = prefix ? `${prefix}.${key}` : key;
-        if (typeof obj[key] === 'object' && obj[key] !== null) {
-            Object.assign(result, flattenMessages(obj[key], fullKey));
-        } else {
-            result[fullKey] = obj[key];
-        }
-    });
-    return result;
-}
-
-const enMessages = flattenMessages(enData.messages);
-const frMessages = flattenMessages(frData.messages);
+const { enMessages, frMessages } = require('../../../../__tests__/fixtures/translations');
 
 const mockStore = {
     getState: () => ({
