+++ geonode_mapstore_client/client/js/plugins/hydrata/Swamps/__tests__/swampsI18n-test.js
@@ -1,21 +1,6 @@
 import expect from 'expect';
 
-const enData = require('../../../../../../static/mapstore/hydrata-translations/data.en-US.json');
-
-function flattenMessages(obj, prefix) {
-    let result = {};
-    Object.keys(obj).forEach(key => {
-        const fullKey = prefix ? `${prefix}.${key}` : key;
-        if (typeof obj[key] === 'object' && obj[key] !== null) {
-            Object.assign(result, flattenMessages(obj[key], fullKey));
-        } else {
-            result[fullKey] = obj[key];
-        }
-    });
-    return result;
-}
-
-const enMessages = flattenMessages(enData.messages);
+const { enMessages } = require('../../../../__tests__/fixtures/translations');
 
 describe('Swamps i18n', () => {
     it('swamps msgIds exist in en-US translation file', () => {
